+++ Extension/src/background/engine/engine-mv2.ts
@@ -66,14 +66,14 @@ export class Engine implements TsWebExtensionEngine {
      */
     async start(): Promise<void> {
         /**
-         * By the rules of Firefox AMO we cannot use remote scripts (and our JS rules can be counted as such).
-         * Because of that we use the following approach (that was accepted by AMO reviewers):
+         * By the rules of Firefox AMO, we cannot use remote scripts (and our JS rules can be counted as such).
+         * Because of that, we use the following approach (that was accepted by AMO reviewers):
          *
          * 1. We pre-build JS rules from AdGuard filters into the JSON file.
-         * 2. At runtime we check every JS rule if it's included into JSON.
-         *  If it is included we allow this rule to work since it's pre-built. Other rules are discarded.
-         * 3. We also allow "User rules" to work since those rules are added manually by the user.
-         *  This way filters maintainers can test new rules before including them in the filters.
+         * 2. At runtime we check every JS rule if it is included into JSON.
+         *    If it is included we allow this rule to work since it is pre-built. Other rules are discarded.
+         * 3. We also allow "User rules" and "Custom filters" to work since those rules are added manually by the user.
+         *    This way filters maintainers can test new rules before including them in the filters.
          */
         if (IS_FIREFOX_AMO) {
             const localScriptRules = await network.getLocalScriptRules();