+++ CHANGELOG.md
@@ -45,6 +45,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Unreleased]: https://github.com/AdguardTeam/AdguardBrowserExtension/compare/v4.4.49...HEAD
 
+
+## [5.0.178] - 2024-12-24
+
+### Changed
+
+- Remade JS rules injections in MV3:
+    - use `chrome.scripting` API for injecting functions for script rules from the pre-built filters,
+    - use script tag injection only for script rules manually added by users —
+      rules from *User rules* and *Custom filters*.
+- Updated [@adguard/tswebextension] to v2.4.0-alpha.8.
+
+### Removed
+
+- Ability to execute new `AG_` script rules from *User rules* and *Custom filters* in MV3.
+
 ## [4.4.48] - 2024-11-25
 
 ### Added
@@ -70,10 +85,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [5.0.170] - 2024-10-30
 
+### Changed
+
+- Updated [@adguard/tswebextension] to v2.4.0-alpha.7.
+
 ### Fixed
 
 - Memory leak caused by multiple script injections on the same pages
-  after a service worker or event page restart in MV3 [#2594].
+  after a service worker or event page restart [#2594].
 
 [5.0.170]: https://github.com/AdguardTeam/AdguardBrowserExtension/releases/tag/v5.0.170
 [#2594]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2594